meson: specify correct minimum versions of Meson
The main project doesn't really have any specific version requirement.
Specify a very low one just in case -- 0.37.0 is old enough to cover
probably any existing use of Meson anywhere in the wild, and
coincidentally is also the version that Meson started adding feature
warnings for, to notify you if you use too-new features.

The example *does* depend on a specific version. It needs 0.55.0 in
order to use subproject wrap dependency fallback instead of the legacy
style of specifying the name of the variable as a fallback. Ensure that
is used.