Don't install anything when building as subproject
When a project is consuming unity as as subproject, unity headers,
static library and pkg config files are installed by `meson install`.

This can be fixed by using `meson install --skip-subprojects`, but this
must be repeated in all the distros packaging a project.

Fixed by disabling install when building as a subproject.

Fixes: #661